Volume Slidebars
Took music out of the two scenes where the player actually plays (AlienShip and EarthShip Scenes). There is a settings button that takes you to two slidebars. One is for music and that seems to work across scenes where the music plays. The other, for SFX, is not working yet. So when the player plays and shoots, the volume of the "pew pew" could be too loud. each sound occurs as the projectile is created dynamically. Trying to figure out how to link that to the slidebar.